+++ src/gallium/drivers/zink/zink_context.h
@@ -76,10 +76,13 @@ void
 zink_fence_wait(struct pipe_context *ctx);
 
 void
-zink_wait_on_batch(struct zink_context *ctx, uint64_t batch_id);
+//zink_wait_on_batch(struct zink_context *ctx, uint64_t batch_id);
+zink_wait_on_batch(struct zink_context *ctx, uint32_t batch_id);
 
 bool
-zink_check_batch_completion(struct zink_context *ctx, uint64_t batch_id);
+//zink_check_batch_completion(struct zink_context *ctx, uint64_t batch_id);
+zink_check_batch_completion(struct zink_context *ctx, uint32_t batch_id);
+
 VkCommandBuffer
 zink_get_cmdbuf(struct zink_context *ctx, struct zink_resource *src, struct zink_resource *dst);
 void
